Rename module file mod.ts
